delete evel submodule
